Stop detecting outlook blockquotes
The id we used to detect outlook blockquotes is not enough. The div with this id only contains the header of the reply/forward but the content of the body of the previous email is right next to this div, at the same level as the div and was therefore not detected.

Instead of half detecting outlook blockquotes, don't detect anything for now until we truly detect them correctly