Protect against empty isThemeTheSameMap and add Sentry to better understand the issue
Right now we have a deeper issue with recycling of Threads or with restPanes() on tablet in landscape. This issue probably causes even more issues in the app that do not crash the app which makes it harder to detect.

This PR at least avoids the crash and tries to gather more info on the source of the issue but the problem persists and needs to be fixed in another PR