Fix missing selection indicator handles when the selected line is on the last line of the rich html editor
If there isn't enough space below the last line of the rich html editor, the selection handles do not show when selecting the last line. The handles should be drawn below the last line but apparently, not having enough space makes them not start drawing at all.

To fix this issue, I always add margins to the bottom of the editor in order to always have exactly one line worth of extra space at the bottom of the editor. To accommodate for the extra space in the editor, I need to add negative margin to the signature so it still looks exactly how it will be when the draft is saved/sent and I need to fix the alignement of the placeholder when the editor is empty.