Delete draft that cannot ever be sent
Most of the sentry issues we receive regarding corrupted attachments come from drafts stored in DB that are retried everytime the worker is started. This bug was caused by the UploadStatus crash inside addDraftBreadcrumbs() for certain. I suspect that there might remain other reasons for the corrupted attachment bug because even with the UploadStatus' runCatching() I did reproduce the corrupted attachement bug once. Also the issue did exist before the addDraftBreadcrumbs() addition and hasn't yet been fixed.

The plan of this commit is to remove the drafts that will never be processed so they stop spamming Sentry. Once this goes live we can see the errors where the attachement is corrupted for another reason than UploadStatus and fix this as well