wip Protect against setting property value that doesn't exist anymore
When a user has a realm migration but comes from a schema version where a value didn't exist, trying to read this value inside realm will throw an error.

During a migration, if we need to write into a property of the new object that has since then disappeared from the model, realm will also throw. It will happen if we have an old migration that initialized old fields that don't exist anymore, for exemple, which is quiet a common situation.