fix: Unsnoozed date display when device's date is in the past
If the device's date was in the past, the unsnoozed date might have been displayed using the relative formatting instead of the default one. This fixes the issue forever.

This occurs if unsnoozed message's date is later than the device's date, which is possible if the device's date is in the past.