fix: When fetching more than one search page, add new page to existing threads
We were replacing all of the existing threads with the new page instead of merging them together which lead to some threads having no more parent folders. This occurred only for threads that were not already downloaded in the app prior to searching. Other threads would have lost the search folder as a parent folder but they still had their original folder which hid the issue in this case