refactor: Remove unreachable condition and avoid main view model dependence
Inside the main view model's `getActionFolderRole()` method, the condition `threads?.firstOrNull()?.folderId == FolderController.SEARCH_FOLDER_ID` can never be met because `folderId` comes from the API and therefore cannot ever be equal to our local `FolderController.SEARCH_FOLDER_ID`. I removed the check, which renders the logic of the method pretty straight forward.

The new logic is so straight forward that, in a lot of cases, there's no need to depend on the view model for this logic. It even introduces unnecessary cognitive complexity to be able to make sense of what is really computed. I kept all the calls centralized in a unique point by introducing the `Thread.folderRole` property