fix: Remove displayBatchedMessages to fix ThreadAdapter
We can't use the properties of an adapter anymore if we use this batched messages strategy. The diff util doesn't see previous instances of the items past the first item because when the 10th item of a list is updated, we've tried binding the 9 previous items one by one just so the 10th could be updated. At the time we try to bind the 10th item again, we can't compare the item with its previous instance because at the last known bind, there was no 10th item there only was 9.

This prevents us from using any form of payload to optimize subsequent binds and avoid binding webviews again which is required because binding webviews again makes all of the ui blink uncontrollably