feat: Remove useless nullability in containsTheSameEmojiValuesAs and fix tests accordingly
containsTheSameEmojiValuesAs() used to require a nullable Reaction inside of the argument maps because of RealmDictionary but it's not the case anymore now that we stopped using RealmDictionary in this situation.

Also fixed the emojiReactionStateOf() instantiation method in the test to use the actual same class used in the app's code. containsTheSameEmojiValuesAs() will compare instances of the interface Reaction using their respective equals() method. To make sure the equals() method is correctly defined for all instances that can be used with containsTheSameEmojiValuesAs() we need to check the same class which is used in the code