fix: Wrap MailBottomSheetScaffoldComposeView into a FrameLayout to avoid animation crash
The system seems to animate these custom views in some rare occasions which causes the TransitionManager to try and call addView() on the root View of the custom view. When using AbstractComposeView as the root view, calls to addView() crash because it's unsupported. To fix the issue, the root view was changed to a view that supports addView().

Fixes MAIL-ANDROID-P6M